Relocate EKS models
